Add functionality to set user and home directory.
The latest version of `omxplayer` stores the D-Bus connection path and
PID in a temp file that has the executing user's name in the filename.

This allows the user of the library to specify what user `omxplayer`
will be running as in case the `USER` and `HOME` environment variables
are not available.